requirements.txt
celery==5.2.7
Django==4.2.16
django-allauth==0.54.0
django-baton==2.7.0
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-cors-headers==3.14.0
django-easy-select2==1.5.8
django-extensions==3.2.1
django-filter==23.1
django-fsm==2.8.1
django-fsm-admin==1.2.5
django-hijack==3.3.0
django-import-export==3.2.0
django-model-utils==4.3.1
django-nepali-datetime-field==0.6
django-object-actions==4.1.0
django-translation-manager==1.3.0
djangorestframework==3.14.0
django-reversion==5.0.4
django_jsignature3==0.9.0
dj-rest-auth==3.0.0
drf-spectacular==0.26.2
fcm-django==1.0.12
Pillow==9.5.0
python-dateutil==2.8.2
python-decouple==3.8
redis==4.5.4
requests==2.28.2
yaspin==2.3.0
django-ckeditor==6.5.1






# Dev requirements
flake8==6.0.0
icecream==2.1.3
isort==5.12.0
pre-commit==3.2.2
yapf==0.33.0
pylint-django==2.5.3
bandit==1.7.5
django-debug-toolbar==4.0.0
